Refactor to use one way to create the SSLContext
