+++ services/gundeck/src/Gundeck/Schema/Run.hs
@@ -17,6 +17,7 @@
 
 module Gundeck.Schema.Run where
 
+import Cassandra.MigrateSchema (migrateSchema)
 import Cassandra.Schema
 import Control.Exception (finally)
 import Gundeck.Schema.V1 qualified as V1
